refactor: implement argparse for user input
Refactor to utilize argparse to close #149. Include a smattering of
tests to ensure the argparse functionality works as expected. Does not
test the functionality of the larger app itself, just the argparse
and sending the correct arguments to global_config. Add a root
logger (which should be customized) and updates the `.toml` file to
include dev dependencies for `pytest` and `ruff`. This likely
deprecates a number of custom functions, so running some static
analysis to find dead code may be beneficial.