refactor: change print_debug to use logging.debug
Deprecate the print_debug function and replaces it with a
logging.debug call. Update readme to document how to use logging
based on using as a standalone app or calling as a dependency.
Add a configuration file for logging which for the time being the user
has to set themselves. Update toml file to point to the correct
function if the user wants to use the woodwork command line tool so
it correctly instantiates its own logging configuration.